requirements.txt
fastapi
uvicorn
python-dotenv
langchain_community
langchain_text_splitters
langchain_core
langchain_chroma
langchain_ollama
python-multipart
pypdf